Add option to start list on same line
Description:
This commit adds the option to start a list on the same line, as
proposed in issue #8. By this, a wrapped list would become:

    void foo(var1,
             var2,
             var3)

instead of:

    void foo(
        var1,
        var2,
        var3
    )

For now, the option is implemented using

    let g:argwrap_wrap_closing_brace = 0

The configuration can be extended to see this option as seperate.

Warning!
It assumes the indentation is done with spaces instead of tabs

Relevant issue: #8 (https://github.com/FooSoft/vim-argwrap/issues/8)